# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nexo API",
        "description": "Fiscal centralizing API for integration between sales systems, government and accounting platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "https://nexo.test",
            "description": "Herd Local Server (HTTPS)"
        },
        {
            "url": "https://backendnexo.shop",
            "description": "Mounted Server"
        }
    ],
    "paths": {
...
        "/api/sales/{saleId}/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "List payments of a sale",
                "description": "Returns the paginated payments applied to the given sale. Admin endpoint, no authentication required.",
                "operationId": "indexPayments",
                "parameters": [
                    {
                        "name": "saleId",
                        "in": "path",
                        "description": "ID of the sale.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of results per page. Defaults to 15.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15,
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payments retrieved successfully."
                    },
                    "404": {
                        "description": "Sale not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a payment for a sale",
                "description": "Persists a new payment snapshot linked to the given sale. Admin endpoint, no authentication required.",
                "operationId": "storePayment",
                "parameters": [
                    {
                        "name": "saleId",
                        "in": "path",
                        "description": "ID of the sale.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "payment_type"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "payment_type": {
                                        "description": "0=Other, 1=Cash, 2=Card, 3=Check, 4=WireTransfer, 5=Voucher, 6=MobileMoney",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created successfully."
                    },
                    "404": {
                        "description": "Sale not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get a payment by ID",
                "description": "Returns a single payment snapshot. Admin endpoint, no authentication required.",
                "operationId": "showPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment retrieved successfully."
                    },
                    "404": {
                        "description": "Payment not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update a payment",
                "description": "Updates amount and/or payment_type. Admin endpoint, no authentication required.",
                "operationId": "updatePayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment to update.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "payment_type": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated successfully."
                    },
                    "404": {
                        "description": "Payment not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete a payment",
                "description": "Permanently removes a payment snapshot. Admin endpoint, no authentication required.",
                "operationId": "destroyPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment to delete.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment deleted successfully."
                    },
                    "404": {
                        "description": "Payment not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "agentToken": {
                "type": "http",
                "description": "Bearer token issued by POST /api/agent/token. Used by the nexo-agent daemon to authenticate all agent endpoints.",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer JWT issued by POST /api/auth/login. Pass as: Authorization: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "connectorToken": {
                "type": "http",
                "description": "Bearer token issued when a Connector is created. Only visible at creation time.",
                "scheme": "bearer"
            }
        }
    },
...
}